requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-AlchemyDumps==0.0.10
flask-marshmallow==0.9.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.1
marshmallow-sqlalchemy==0.13.2
mysqlclient==1.3.12
pycryptodome==3.6.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
python-envcfg==0.2.0
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
Unipath==1.1
urllib3==1.22
Werkzeug==0.14.1
